+++ apps/node/lodge/src/store/game/card-value.ts
@@ -1,24 +1,21 @@
 import { atomFamily, transaction } from "atom.io"
 
 import { createMutableAtom } from "atom.io/internal"
-import { TransceiverSet } from "atom.io/transceivers/set-io"
 import type { Identified } from "~/packages/anvl/src/id"
 import type { Json } from "~/packages/anvl/src/json"
+import { SetRTX } from "~/packages/atom.io/transceivers/set-rtx/src"
 import { AtomicJunction } from "../utils/atomic-junction"
 
 export const findCardValueState = atomFamily<Identified & Json.Object, string>({
 	key: `findCardValue`,
 	default: () => ({ id: `` }),
 })
-export const cardValuesIndex = createMutableAtom<
-	TransceiverSet<string>,
-	string[]
->({
+export const cardValuesIndex = createMutableAtom<SetRTX<string>, string[]>({
 	key: `cardValuesIndex::mutable`,
 	mutable: true,
-	default: () => new TransceiverSet<string>(),
+	default: () => new SetRTX<string>(),
 	toJson: (set) => [...set],
-	fromJson: (array) => new TransceiverSet<string>(array),
+	fromJson: (array) => new SetRTX<string>(array),
 })
 
 export const valuesOfCards = new AtomicJunction({