+++ packages/atom.io/__test__/logger.test.ts
@@ -1,10 +1,10 @@
 import { __INTERNAL__, setLogLevel, useLogger } from "../src"
 
-const doNothing = () => null
-
 describe(`setLogLevel`, () => {
 	it(`allows logging at the preferred level`, () => {
-		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.info).toBe(doNothing)
+		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.info).toBeInstanceOf(
+			Function,
+		)
 		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.warn).toBe(console.warn)
 		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.error).toBe(console.error)
 
@@ -15,13 +15,19 @@ describe(`setLogLevel`, () => {
 		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.error).toBe(console.error)
 
 		setLogLevel(`warn`)
-		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.info).toBe(doNothing)
+		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.info).toBeInstanceOf(
+			Function,
+		)
 		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.warn).toBe(console.warn)
 		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.error).toBe(console.error)
 
 		setLogLevel(`error`)
-		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.info).toBe(doNothing)
-		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.warn).toBe(doNothing)
+		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.info).toBeInstanceOf(
+			Function,
+		)
+		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.warn).toBeInstanceOf(
+			Function,
+		)
 		expect(__INTERNAL__.IMPLICIT.STORE.config.logger?.error).toBe(console.error)
 
 		setLogLevel(null)