+++ packages/atom.io/__test__/mutability/tracker.test.ts
@@ -11,7 +11,7 @@ import {
 	transaction,
 } from "atom.io"
 import { FamilyTracker, Tracker } from "atom.io/internal"
-import { TransceiverSet } from "atom.io/transceivers/set-io"
+import { SetRTX } from "~/packages/atom.io/transceivers/set-rtx/src"
 import * as UTIL from "../__util__"
 
 const LOG_LEVELS = [null, `error`, `warn`, `info`] as const
@@ -29,26 +29,26 @@ beforeEach(() => {
 
 describe(`tracker`, () => {
 	test(`tracks the state of a mutable atom`, () => {
-		const mutableSetState = atom<TransceiverSet<string>>({
+		const mutableSetState = atom<SetRTX<string>>({
 			key: `mutableSetState`,
-			default: new TransceiverSet(),
+			default: new SetRTX(),
 		})
 		const { latestUpdateState } = new Tracker(mutableSetState)
 
-		expect(getState(mutableSetState)).toEqual(new TransceiverSet())
+		expect(getState(mutableSetState)).toEqual(new SetRTX())
 		expect(getState(latestUpdateState)).toEqual(null)
 		setState(latestUpdateState, `0=add:"x"`)
 		expect(getState(latestUpdateState)).toEqual(`0=add:"x"`)
-		expect(getState(mutableSetState)).toEqual(new TransceiverSet([`x`]))
+		expect(getState(mutableSetState)).toEqual(new SetRTX([`x`]))
 		setState(latestUpdateState, `1=add:"y"`)
 		expect(getState(latestUpdateState)).toEqual(`1=add:"y"`)
-		expect(getState(mutableSetState)).toEqual(new TransceiverSet([`x`, `y`]))
+		expect(getState(mutableSetState)).toEqual(new SetRTX([`x`, `y`]))
 	})
 
 	test(`updates its core in a transaction`, () => {
-		const mutableSetState = atom<TransceiverSet<string>>({
+		const mutableSetState = atom<SetRTX<string>>({
 			key: `mutableSetState`,
-			default: new TransceiverSet(),
+			default: new SetRTX(),
 		})
 		const tracker = new Tracker(mutableSetState)
 		const updateTrackerTX = transaction({
@@ -59,28 +59,28 @@ describe(`tracker`, () => {
 			},
 		})
 
-		expect(getState(mutableSetState)).toEqual(new TransceiverSet())
+		expect(getState(mutableSetState)).toEqual(new SetRTX())
 		expect(getState(tracker.latestUpdateState)).toEqual(null)
 		runTransaction(updateTrackerTX)()
-		expect(getState(mutableSetState)).toEqual(new TransceiverSet([`x`, `y`]))
+		expect(getState(mutableSetState)).toEqual(new SetRTX([`x`, `y`]))
 	})
 })
 
 describe(`trackerFamily`, () => {
 	test(`tracks the state of a family of mutable atoms`, () => {
-		const findSetState = atomFamily<TransceiverSet<string>, string>({
+		const findSetState = atomFamily<SetRTX<string>, string>({
 			key: `findSetState`,
-			default: () => new TransceiverSet(),
+			default: () => new SetRTX(),
 		})
 		const { findLatestUpdateState } = new FamilyTracker(findSetState)
 
-		expect(getState(findSetState(`a`))).toEqual(new TransceiverSet())
+		expect(getState(findSetState(`a`))).toEqual(new SetRTX())
 		expect(getState(findLatestUpdateState(`a`))).toEqual(null)
 	})
 	test(`updates the core of a new family member in a transaction`, () => {
-		const findSetState = atomFamily<TransceiverSet<string>, string>({
+		const findSetState = atomFamily<SetRTX<string>, string>({
 			key: `findSetState`,
-			default: () => new TransceiverSet(),
+			default: () => new SetRTX(),
 		})
 		const { findLatestUpdateState } = new FamilyTracker(findSetState)
 		const updateTrackerTX = transaction<(key: string) => void>({
@@ -91,11 +91,11 @@ describe(`trackerFamily`, () => {
 			},
 		})
 
-		expect(getState(findSetState(`a`))).toEqual(new TransceiverSet())
+		expect(getState(findSetState(`a`))).toEqual(new SetRTX())
 		expect(getState(findLatestUpdateState(`a`))).toEqual(null)
 		runTransaction(updateTrackerTX)(`a`)
 
-		expect(getState(findSetState(`a`))).toEqual(new TransceiverSet([`x`]))
-		expect(getState(findSetState(`b`))).toEqual(new TransceiverSet())
+		expect(getState(findSetState(`a`))).toEqual(new SetRTX([`x`]))
+		expect(getState(findSetState(`b`))).toEqual(new SetRTX())
 	})
 })