+++ packages/atom.io/json/src/index.ts
@@ -1,4 +1,5 @@
-export * from "anvl/json"
+export * from "~/packages/anvl/json"
+export * from "~/packages/anvl/src/primitive"
 
 export * from "./select-json"
 export * from "./select-json-family"