+++ packages/atom.io/tsup.config.ts
@@ -7,6 +7,5 @@ export default defineConfig({
 	splitting: true,
 	treeshake: true,
 	sourcemap: true,
-	minify: true,
 	clean: true,
 })